użytkownicy, którzy dodali swoją firmę, mogą ją edytować lub usuwać

+++ CarsAPI/Services/CarCompanyService.cs
@@ -1,31 +1,37 @@
 ﻿using AutoMapper;
+using CarsAPI.Authorization;
 using CarsAPI.Entities;
+using CarsAPI.Exceptions;
 using CarsAPI.Models;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.EntityFrameworkCore;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 
 namespace CarsAPI.Services
 {
     public interface ICarCompanyService
     {
-        bool DeletedCompany(int id);
+        bool DeletedCompany(int id, ClaimsPrincipal user);
         IEnumerable<CarCompanyDto> GetCompanies();
         CarCompanyDto GetCompany(int id);
-        int NewCompany(CreateCarCompanyDto dto);
-        bool Update(CreateCarCompanyDto dto, int id);
+        int NewCompany(CreateCarCompanyDto dto, int userId);
+        bool Update(CreateCarCompanyDto dto, int id, ClaimsPrincipal user);
     }
     public class CarCompanyService : ICarCompanyService
     {
         private readonly CarDbContext context;
         private readonly IMapper mapper;
+        private readonly IAuthorizationService authorizationService;
 
-        public CarCompanyService(CarDbContext context, IMapper mapper)
+        public CarCompanyService(CarDbContext context, IMapper mapper, IAuthorizationService authorizationService)
         {
             this.context = context;
             this.mapper = mapper;
+            this.authorizationService = authorizationService;
         }
         public CarCompanyDto GetCompany(int id)
         {
@@ -49,9 +55,10 @@ namespace CarsAPI.Services
             return companiesDtos;
         }
 
-        public int NewCompany(CreateCarCompanyDto dto)
+        public int NewCompany(CreateCarCompanyDto dto, int userId)
         {
             var company = mapper.Map<CarCompany>(dto);
+            company.CreateById = userId;
 
             context.CarCompany.Add(company);
             context.SaveChanges();
@@ -60,12 +67,20 @@ namespace CarsAPI.Services
 
         }
 
-        public bool DeletedCompany(int id)
+        public bool DeletedCompany(int id, ClaimsPrincipal user)
         {
             var deleted = context.CarCompany.FirstOrDefault(c => c.Id == id);
 
             if (deleted is null)
                 return false;
+
+            var authorizationResult = authorizationService.AuthorizeAsync(user, deleted, new ResourceOperationRequirement(ResourceOperation.Delete
+                )).Result;
+
+            if (!authorizationResult.Succeeded)
+            {
+                throw new ForbidException();
+            }
 
             context.CarCompany.RemoveRange(deleted);
             context.SaveChanges();
@@ -73,12 +88,19 @@ namespace CarsAPI.Services
             return true;
         }
 
-        public bool Update(CreateCarCompanyDto dto, int id)
+        public bool Update(CreateCarCompanyDto dto, int id, ClaimsPrincipal user)
         {
             var updated = context.CarCompany.FirstOrDefault(c => c.Id == id);
 
             if (updated is null)
                 return false;
+
+            var authorizationResult = authorizationService.AuthorizeAsync(user, updated, new ResourceOperationRequirement(ResourceOperation.Update)).Result;
+
+            if (!authorizationResult.Succeeded)
+            {
+                throw new ForbidException();
+            }
 
             updated.LegalForm = dto.LegalForm;
             updated.Name = dto.Name;
